pyproject.toml
[tool.poetry]
name = "dbt-python-snowpark-f1"
version = "0.1.0"
description = ""
authors = ["Greg Clunies <greg.clunies@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10,<3.12"
dbt-snowflake = "^1.7.2"
snowflake-snowpark-python = "^1.12.1"
python-dotenv = "^1.0.1"
ipykernel = "^6.29.2"
snowflake-ml-python = "^1.2.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
